requirements.txt
streamlit
opencv-python-headless
numpy
pillow
python-dotenv
google-generativeai
fastapi
uvicorn
python-multipart
